Fetch user names (not just IDs)
Any files archived will need a human-friendly filename that includes
the name of the user ("joel"), not just the ID ("CIHEO3FFJ"). Very
similar to what we just did with channels in c434141.
`channel_list_request` gets generalized to handle user lists as well
since the requests are about identical in format, and we make a dict
from the response just as with channels.