Add 'action' attribute to SlackFile
The script originally just deleted files, but now we’re going to be
taking possibly multiple actions on each file (archiving, etc). So
instead of filtering the collection of files to the set we want and
taking a single action on all of them, we will be shifting to an
approach of marking each file with the action we intend to take. This
way it can be logged along with everything else about the file, and we
can wash through the collection multiple times: archive all the ones
marked “archive”, then “delete” etc.